feat(dashboard): renaming former dashboard to bonding
