+++ components/AssetInput/AssetList.tsx
@@ -2,8 +2,8 @@ import { FC, useMemo } from 'react'
 
 import { Box, Button, HStack, Image, Text } from '@chakra-ui/react'
 import FallbackImage from 'components/FallbackImage'
-import { TokenBalance } from 'components/Pages/Dashboard/BondingActions/Bond'
-import { useConfig } from 'components/Pages/Dashboard/hooks/useDashboardData'
+import { TokenBalance } from 'components/Pages/Bonding/BondingActions/Bond'
+import { useConfig } from 'components/Pages/Bonding/hooks/useDashboardData'
 import useFilter from 'hooks/useFilter'
 import { useMultipleTokenBalance } from 'hooks/useTokenBalance'
 import { useTokenList } from 'hooks/useTokenList'
