+++ components/Pages/Bonding/BondingActions/BondingActions.tsx
@@ -4,19 +4,19 @@ import { ArrowBackIcon } from '@chakra-ui/icons'
 import { Box, Button, HStack, IconButton, Text, VStack } from '@chakra-ui/react'
 import { useChain } from '@cosmos-kit/react-lite'
 import Loader from 'components/Loader'
-import { Bond, BondingTokenState } from 'components/Pages/Dashboard/BondingActions/Bond'
-import { bondingAtom } from 'components/Pages/Dashboard/BondingActions/bondAtoms'
-import { BondingActionTooltip } from 'components/Pages/Dashboard/BondingActions/BondingActionTooltip'
-import useTransaction from 'components/Pages/Dashboard/BondingActions/hooks/useTransaction'
-import Unbond from 'components/Pages/Dashboard/BondingActions/Unbond'
-import Withdraw from 'components/Pages/Dashboard/BondingActions/Withdraw'
-import { ActionType } from 'components/Pages/Dashboard/BondingOverview'
-import { WithdrawableInfo } from 'components/Pages/Dashboard/hooks/getWithdrawable'
+import { Bond, BondingTokenState } from 'components/Pages/Bonding/BondingActions/Bond'
+import { bondingAtom } from 'components/Pages/Bonding/BondingActions/bondAtoms'
+import { BondingActionTooltip } from 'components/Pages/Bonding/BondingActions/BondingActionTooltip'
+import useTransaction from 'components/Pages/Bonding/BondingActions/hooks/useTransaction'
+import Unbond from 'components/Pages/Bonding/BondingActions/Unbond'
+import Withdraw from 'components/Pages/Bonding/BondingActions/Withdraw'
+import { ActionType } from 'components/Pages/Bonding/BondingOverview'
+import { WithdrawableInfo } from 'components/Pages/Bonding/hooks/getWithdrawable'
 import {
   Config,
   useConfig,
   useDashboardData,
-} from 'components/Pages/Dashboard/hooks/useDashboardData'
+} from 'components/Pages/Bonding/hooks/useDashboardData'
 import { kBg, kBorderRadius } from 'constants/visualComponentConstants'
 import { useMultipleTokenBalance } from 'hooks/useTokenBalance'
 import { useRouter } from 'next/router'
