+++ components/Pages/Bonding/BondingActions/Withdraw.tsx
@@ -8,11 +8,11 @@ import {
   VStack,
 } from '@chakra-ui/react'
 import { useChain } from '@cosmos-kit/react-lite'
-import { BondingTokenState } from 'components/Pages/Dashboard/BondingActions/Bond'
-import { UnbondingData } from 'components/Pages/Dashboard/hooks/getUnbonding'
-import { WithdrawableInfo } from 'components/Pages/Dashboard/hooks/getWithdrawable'
-import { useConfig } from 'components/Pages/Dashboard/hooks/useDashboardData'
-import { WhaleTooltip } from 'components/Pages/Dashboard/WhaleTooltip'
+import { BondingTokenState } from 'components/Pages/Bonding/BondingActions/Bond'
+import { UnbondingData } from 'components/Pages/Bonding/hooks/getUnbonding'
+import { WithdrawableInfo } from 'components/Pages/Bonding/hooks/getWithdrawable'
+import { useConfig } from 'components/Pages/Bonding/hooks/useDashboardData'
+import { WhaleTooltip } from 'components/Pages/Bonding/WhaleTooltip'
 import { usePrices } from 'hooks/usePrices'
 import { useTokenList } from 'hooks/useTokenList'
 import { useRecoilState, useRecoilValue } from 'recoil'
