+++ components/Pages/Bonding/BondingActions/hooks/bondTokens.ts
@@ -1,32 +1,43 @@
 import { SigningCosmWasmClient } from '@cosmjs/cosmwasm-stargate/build/signingcosmwasmclient'
+import { coin } from '@cosmjs/stargate'
 import { InjectiveSigningStargateClient } from '@injectivelabs/sdk-ts/dist/cjs/core/stargate'
-import { Config } from 'components/Pages/Dashboard/hooks/useDashboardData'
+import { Config } from 'components/Pages/Bonding/hooks/useDashboardData'
 import { ChainId } from 'constants/index'
 import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 import { TerraTreasuryService } from 'services/treasuryService'
 import { getInjectiveTxData } from 'util/injective'
 import { createExecuteMessage } from 'util/messages/createExecuteMessage'
 
-export const createNewEpoch: any = async (
+export const bondTokens: any = async (
   signingClient: SigningCosmWasmClient,
-  config: Config,
   address: string,
+  amount: number,
+  denom: string,
+  config: Config,
   injectiveSigningClient?: InjectiveSigningStargateClient
 ) => {
   const handleMsg = {
-    new_epoch: {},
+    bond: {
+      asset: {
+        amount: amount.toString(),
+        info: {
+          native_token: {
+            denom,
+          },
+        },
+      },
+    },
   }
-
   const execMsg = createExecuteMessage({ senderAddress: address,
-    contractAddress: config.fee_distributor,
+    contractAddress: config.whale_lair,
     message: handleMsg,
-    funds: [] })
+    funds: [coin(amount, denom)] })
   let fee: any = 'auto'
   if (await signingClient.getChainId() === ChainId.terrac) {
     const gas = Math.ceil(await signingClient.simulate(
       address, [execMsg], '',
     ) * 1.3)
-    fee = await TerraTreasuryService.getInstance().getTerraClassicFee(null, gas)
+    fee = await TerraTreasuryService.getInstance().getTerraClassicFee(execMsg.value.funds, gas)
   } else if (injectiveSigningClient && await signingClient.getChainId() === ChainId.injective) {
     const injectiveTxData = await getInjectiveTxData(
       injectiveSigningClient, address, [execMsg],
@@ -37,3 +48,4 @@ export const createNewEpoch: any = async (
     address, [execMsg], fee, '',
   )
 }
+
