+++ components/Pages/Bonding/BondingActions/hooks/claimRewards.ts
@@ -1,43 +1,32 @@
 import { SigningCosmWasmClient } from '@cosmjs/cosmwasm-stargate/build/signingcosmwasmclient'
-import { coin } from '@cosmjs/stargate'
-import { InjectiveSigningStargateClient } from '@injectivelabs/sdk-ts/dist/cjs/core/stargate'
-import { Config } from 'components/Pages/Dashboard/hooks/useDashboardData'
+import { InjectiveSigningStargateClient } from '@injectivelabs/sdk-ts/dist/cjs/core/stargate';
+import { Config } from 'components/Pages/Bonding/hooks/useDashboardData'
 import { ChainId } from 'constants/index'
 import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 import { TerraTreasuryService } from 'services/treasuryService'
 import { getInjectiveTxData } from 'util/injective'
 import { createExecuteMessage } from 'util/messages/createExecuteMessage'
 
-export const bondTokens: any = async (
+export const claimRewards: any = async (
   signingClient: SigningCosmWasmClient,
   address: string,
-  amount: number,
-  denom: string,
   config: Config,
   injectiveSigningClient?: InjectiveSigningStargateClient
 ) => {
   const handleMsg = {
-    bond: {
-      asset: {
-        amount: amount.toString(),
-        info: {
-          native_token: {
-            denom,
-          },
-        },
-      },
-    },
+    claim: {},
   }
+
   const execMsg = createExecuteMessage({ senderAddress: address,
-    contractAddress: config.whale_lair,
+    contractAddress: config.fee_distributor,
     message: handleMsg,
-    funds: [coin(amount, denom)] })
+    funds: [] })
   let fee: any = 'auto'
   if (await signingClient.getChainId() === ChainId.terrac) {
     const gas = Math.ceil(await signingClient.simulate(
       address, [execMsg], '',
     ) * 1.3)
-    fee = await TerraTreasuryService.getInstance().getTerraClassicFee(execMsg.value.funds, gas)
+    fee = await TerraTreasuryService.getInstance().getTerraClassicFee(null, gas)
   } else if (injectiveSigningClient && await signingClient.getChainId() === ChainId.injective) {
     const injectiveTxData = await getInjectiveTxData(
       injectiveSigningClient, address, [execMsg],
@@ -48,4 +37,3 @@ export const bondTokens: any = async (
     address, [execMsg], fee, '',
   )
 }
-
