+++ components/Pages/Bonding/BondingActions/hooks/useTransaction.tsx
@@ -4,16 +4,16 @@ import { useMutation, useQuery } from 'react-query'
 import { useToast } from '@chakra-ui/react'
 import { useChain } from '@cosmos-kit/react-lite'
 import Finder from 'components/Finder'
-import { bondTokens } from 'components/Pages/Dashboard/BondingActions/hooks/bondTokens'
-import { claimRewards } from 'components/Pages/Dashboard/BondingActions/hooks/claimRewards'
-import { createNewEpoch } from 'components/Pages/Dashboard/BondingActions/hooks/createNewEpoch'
-import { unbondTokens } from 'components/Pages/Dashboard/BondingActions/hooks/unbondTokens'
-import { withdrawTokens } from 'components/Pages/Dashboard/BondingActions/hooks/withdrawTokens'
-import { ActionType } from 'components/Pages/Dashboard/BondingOverview'
+import { bondTokens } from 'components/Pages/Bonding/BondingActions/hooks/bondTokens'
+import { claimRewards } from 'components/Pages/Bonding/BondingActions/hooks/claimRewards'
+import { createNewEpoch } from 'components/Pages/Bonding/BondingActions/hooks/createNewEpoch'
+import { unbondTokens } from 'components/Pages/Bonding/BondingActions/hooks/unbondTokens'
+import { withdrawTokens } from 'components/Pages/Bonding/BondingActions/hooks/withdrawTokens'
+import { ActionType } from 'components/Pages/Bonding/BondingOverview'
 import {
   Config,
   useConfig,
-} from 'components/Pages/Dashboard/hooks/useDashboardData'
+} from 'components/Pages/Bonding/hooks/useDashboardData'
 import { useClients } from 'hooks/useClients'
 import { useRecoilValue } from 'recoil'
 import { chainState } from 'state/chainState'
