+++ components/Pages/Bonding/BondingActions/index.ts
@@ -1,0 +1,1 @@
+export { default } from 'components/Pages/Bonding/BondingActions/BondingActions'
