+++ components/Pages/Bonding/Dashboard.tsx
@@ -3,10 +3,10 @@ import { FC, useEffect, useMemo, useState } from 'react'
 import { Flex, HStack, Text, useMediaQuery, VStack } from '@chakra-ui/react'
 import { useChain } from '@cosmos-kit/react-lite'
 import Loader from 'components/Loader'
-import { TokenBalance } from 'components/Pages/Dashboard/BondingActions/Bond'
-import { BondedData } from 'components/Pages/Dashboard/hooks/getBonded'
-import { UnbondingData } from 'components/Pages/Dashboard/hooks/getUnbonding'
-import { WithdrawableInfo } from 'components/Pages/Dashboard/hooks/getWithdrawable'
+import { TokenBalance } from 'components/Pages/Bonding/BondingActions/Bond'
+import { BondedData } from 'components/Pages/Bonding/hooks/getBonded'
+import { UnbondingData } from 'components/Pages/Bonding/hooks/getUnbonding'
+import { WithdrawableInfo } from 'components/Pages/Bonding/hooks/getWithdrawable'
 import { WHALE_TOKEN_SYMBOL } from 'constants/index'
 import { usePrices } from 'hooks/usePrices'
 import { useMultipleTokenBalance } from 'hooks/useTokenBalance'
