+++ components/Pages/Bonding/hooks/getGlobalIndex.ts
