+++ components/Pages/Bonding/hooks/getTotalBonded.ts
