+++ components/Pages/Bonding/hooks/getWeight.ts
