+++ components/Pages/Bonding/hooks/useDashboardData.ts
@@ -1,16 +1,16 @@
 import { useEffect, useMemo, useState } from 'react'
 import { useQueries } from 'react-query'
 
-import { getBonded } from 'components/Pages/Dashboard/hooks/getBonded'
-import { getBondingConfig } from 'components/Pages/Dashboard/hooks/getBondingConfig'
-import { getClaimable } from 'components/Pages/Dashboard/hooks/getClaimable'
-import { getClaimableEpochs } from 'components/Pages/Dashboard/hooks/getClaimableEpochs'
-import { getCurrentEpoch } from 'components/Pages/Dashboard/hooks/getCurrentEpoch'
-import { getFeeDistributorConfig } from 'components/Pages/Dashboard/hooks/getFeeDistributorConfig'
-import { getTotalBonded } from 'components/Pages/Dashboard/hooks/getTotalBonded'
-import { getUnbonding } from 'components/Pages/Dashboard/hooks/getUnbonding'
-import { getWeight } from 'components/Pages/Dashboard/hooks/getWeight'
-import { getWithdrawable } from 'components/Pages/Dashboard/hooks/getWithdrawable'
+import { getBonded } from 'components/Pages/Bonding/hooks/getBonded'
+import { getBondingConfig } from 'components/Pages/Bonding/hooks/getBondingConfig'
+import { getClaimable } from 'components/Pages/Bonding/hooks/getClaimable'
+import { getClaimableEpochs } from 'components/Pages/Bonding/hooks/getClaimableEpochs'
+import { getCurrentEpoch } from 'components/Pages/Bonding/hooks/getCurrentEpoch'
+import { getFeeDistributorConfig } from 'components/Pages/Bonding/hooks/getFeeDistributorConfig'
+import { getTotalBonded } from 'components/Pages/Bonding/hooks/getTotalBonded'
+import { getUnbonding } from 'components/Pages/Bonding/hooks/getUnbonding'
+import { getWeight } from 'components/Pages/Bonding/hooks/getWeight'
+import { getWithdrawable } from 'components/Pages/Bonding/hooks/getWithdrawable'
 import { DEFAULT_TOKEN_BALANCE_REFETCH_INTERVAL } from 'constants/settings'
 import { useClients } from 'hooks/useClients'
 import { debounce } from 'lodash'
