+++ components/Pages/Bonding/index.ts
