--- components/Pages/Dashboard/BondingActions/index.ts
@@ -1,1 +1,0 @@
-export { default } from 'components/Pages/Dashboard/BondingActions/BondingActions'
