+++ components/Pages/Trade/Incentivize/hooks/useCurrentEpoch.ts
@@ -1,8 +1,8 @@
 import { useQuery } from 'react-query'
 
 import { CosmWasmClient } from '@cosmjs/cosmwasm-stargate'
-import { fetchCurrentEpoch } from 'components/Pages/Dashboard/hooks/getCurrentEpoch'
-import { Config } from 'components/Pages/Dashboard/hooks/useDashboardData'
+import { fetchCurrentEpoch } from 'components/Pages/Bonding/hooks/getCurrentEpoch'
+import { Config } from 'components/Pages/Bonding/hooks/useDashboardData'
 
 export const useCurrentEpoch = (client, config) => useQuery(
   ['useCurrentEpoch'], async () => getCurrentEpoch(client, config), {
