+++ components/Pages/Trade/Incentivize/hooks/useIncentivePoolInfo.ts
@@ -5,7 +5,7 @@ import { CosmWasmClient } from '@cosmjs/cosmwasm-stargate'
 import {
   Config,
   useConfig,
-} from 'components/Pages/Dashboard/hooks/useDashboardData'
+} from 'components/Pages/Bonding/hooks/useDashboardData'
 import { useCurrentEpoch } from 'components/Pages/Trade/Incentivize/hooks/useCurrentEpoch'
 import {
   ChainId,
