+++ components/Pages/Trade/Liquidity/Claim.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 
 import { HStack, VStack } from '@chakra-ui/react'
 import { useChain } from '@cosmos-kit/react-lite'
-import { useConfig } from 'components/Pages/Dashboard/hooks/useDashboardData'
+import { useConfig } from 'components/Pages/Bonding/hooks/useDashboardData'
 import ClaimTable from 'components/Pages/Trade/Liquidity/ClaimTable'
 import { useCheckIncentiveSnapshots } from 'components/Pages/Trade/Liquidity/hooks/useCheckIncentiveSnapshots'
 import { useClaim } from 'components/Pages/Trade/Liquidity/hooks/useClaim'
