+++ components/Pages/Trade/Liquidity/hooks/useCheckIncentiveSnapshots.ts
@@ -1,7 +1,7 @@
 import { useQuery } from 'react-query'
 
 import { CosmWasmClient } from '@cosmjs/cosmwasm-stargate'
-import { Config } from 'components/Pages/Dashboard/hooks/useDashboardData'
+import { Config } from 'components/Pages/Bonding/hooks/useDashboardData'
 import { useCurrentEpoch } from 'components/Pages/Trade/Incentivize/hooks/useCurrentEpoch'
 import { useQueryIncentiveContracts } from 'components/Pages/Trade/Incentivize/hooks/useQueryIncentiveContracts'
