+++ components/Pages/Trade/Liquidity/hooks/useForceEpochAndTakingSnapshots.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 import { useMutation } from 'react-query'
 
 import { useChain } from '@cosmos-kit/react-lite'
-import { Config } from 'components/Pages/Dashboard/hooks/useDashboardData'
+import { Config } from 'components/Pages/Bonding/hooks/useDashboardData'
 import { useQueryIncentiveContracts } from 'components/Pages/Trade/Incentivize/hooks/useQueryIncentiveContracts'
 import { ChainId } from 'constants/index'
 import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
