+++ pages/[chainId]/dashboard/unbond.tsx
@@ -1,5 +1,5 @@
-import BondingActions from 'components/Pages/Dashboard/BondingActions'
-import { ActionType } from 'components/Pages/Dashboard/BondingOverview'
+import BondingActions from 'components/Pages/Bonding/BondingActions'
+import { ActionType } from 'components/Pages/Bonding/BondingOverview'
 
 const UnbondPage = () => <BondingActions globalAction={ActionType.unbond} />
